Refactor into `Domain` subnamespace
