+++ spec/Technodelight/Jira/Console/Argument/IssueKeyResolverSpec.php
@@ -3,12 +3,10 @@
 namespace spec\Technodelight\Jira\Console\Argument;
 
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Symfony\Component\Console\Input\InputInterface;
 use Technodelight\Jira\Api\GitShell\Api as Git;
 use Technodelight\Jira\Api\GitShell\Branch;
 use Technodelight\Jira\Configuration\ApplicationConfiguration;
-use Technodelight\Jira\Console\Argument\Exception\MissingIssueKeyException;
 use Technodelight\Jira\Console\Argument\IssueKey;
 use Technodelight\Jira\Console\Argument\IssueKeyResolver;
