+++ spec/Technodelight/Jira/Console/Argument/IssueKeySpec.php
@@ -3,11 +3,9 @@
 namespace spec\Technodelight\Jira\Console\Argument;
 
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Symfony\Component\Console\Input\InputInterface;
 use Technodelight\Jira\Api\GitShell\Branch;
 use Technodelight\Jira\Console\Argument\Exception\MissingIssueKeyException;
-use Technodelight\Jira\Console\Argument\IssueKey;
 
 class IssueKeySpec extends ObjectBehavior
 {
