+++ src/Technodelight/Jira/Domain/IssueCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Technodelight\Jira\Api;
+namespace Technodelight\Jira\Domain;
 
 use Iterator;
 use Countable;
